pkg/cdi: fix a build failure on windows.
Fix the most obvious build failure on windows.
